requirements.txt
tensorflow==2.13.0
pandas==2.1.0
numpy==1.24.3
Pillow==10.0.1
matplotlib==3.8.0
fastapi==0.103.1
python-multipart==0.0.6
uvicorn==0.23.2
